+++ CHANGELOG.md
@@ -12,12 +12,16 @@ Version 56
 ----------
 *in development*
 
+#### Next release
+*TBD*
+
+* 🔧 Upload multiple plays to Trakt during first sync (previously would only upload one).
+
 #### 56-beta4
 *2020-10-09*
 
 * 🔧 Improvements to background tasks, billing.
 * 🔨 Do not crash when loading movie with invalid release date.
-* 🔧 Optimizations for Android 11.
 
 #### 56-beta3
 *2020-10-01*